move saveuserdata from header.js to home.js
